Refactor as an answer to review comments
The changes requested are resolved here.
